+++ packages/svelte/src/useForm.js
@@ -1,12 +1,13 @@
 import { router } from '@inertiajs/core'
 import isEqual from 'lodash.isequal'
+import cloneDeep from 'lodash.clonedeep'
 import { writable } from 'svelte/store'
 
 function useForm(...args) {
   const rememberKey = typeof args[0] === 'string' ? args[0] : null
   const data = (typeof args[0] === 'string' ? args[1] : args[0]) || {}
   const restored = rememberKey ? router.restore(rememberKey) : null
-  let defaults = data
+  let defaults = cloneDeep(data)
   let cancelToken = null
   let recentlySuccessfulTimeoutId = null
   let transform = (data) => data
@@ -22,7 +23,7 @@ function useForm(...args) {
     processing: false,
     setStore(key, value) {
       store.update((store) => {
-        return Object.assign({}, store, typeof key === 'string' ? { [key]: value } : key)
+        return Object.assign(store, typeof key === 'string' ? { [key]: value } : key)
       })
     },
     data() {
@@ -38,24 +39,25 @@ function useForm(...args) {
     },
     defaults(key, value) {
       if (typeof key === 'undefined') {
-        defaults = Object.assign(defaults, this.data())
+        defaults = Object.assign(defaults, cloneDeep(this.data()))
 
         return this
       }
 
-      defaults = Object.assign(defaults, value ? { [key]: value } : key)
+      defaults = Object.assign(defaults, cloneDeep(value ? { [key]: value } : key))
 
       return this
     },
     reset(...fields) {
+      let clonedDefaults = cloneDeep(defaults)
       if (fields.length === 0) {
-        this.setStore(defaults)
+        this.setStore(clonedDefaults)
       } else {
         this.setStore(
-          Object.keys(defaults)
+          Object.keys(clonedDefaults)
             .filter((key) => fields.includes(key))
             .reduce((carry, key) => {
-              carry[key] = defaults[key]
+              carry[key] = clonedDefaults[key]
               return carry
             }, {}),
         )