+++ playgrounds/react/resources/js/app.tsx
@@ -2,6 +2,7 @@ import { createInertiaApp } from '@inertiajs/react'
 import { hydrateRoot } from 'react-dom/client'
 
 createInertiaApp({
+  title: (title) => `${title} - React Playground`,
   resolve: (name) => {
     const pages = import.meta.glob('./Pages/**/*.tsx', { eager: true })
     return pages[`./Pages/${name}.tsx`]