+++ playgrounds/react/resources/js/ssr.tsx
@@ -6,6 +6,7 @@ createServer((page) =>
   createInertiaApp({
     page,
     render: ReactDOMServer.renderToString,
+    title: (title) => `${title} - React Playground`,
     resolve: (name) => {
       const pages = import.meta.glob('./Pages/**/*.tsx', { eager: true })
       return pages[`./Pages/${name}.tsx`]